+++ spec/factories/early_career_payments/eligibilities.rb
@@ -2,6 +2,7 @@ FactoryBot.define do
   factory :early_career_payments_eligibility, class: "EarlyCareerPayments::Eligibility" do
     trait :eligible do
       nqt_in_academic_year_after_itt { true }
+      pgitt_or_ugitt_course { :postgraduate }
     end
   end
 end