+++ spec/models/early_career_payments/eligibility_answers_presenter_spec.rb
@@ -3,7 +3,9 @@ require "rails_helper"
 RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
   let(:eligibility_attributes) do
     {
-      nqt_in_academic_year_after_itt: true
+      nqt_in_academic_year_after_itt: true,
+      pgitt_or_ugitt_course: :postgraduate
+
     }
   end
   let(:eligibility) { claim.eligibility }
@@ -13,7 +15,8 @@ RSpec.describe EarlyCareerPayments::EligibilityAnswersPresenter do
 
   it "returns an array of questions and answers to be presented to the user for checking" do
     expected_answers = [
-      [I18n.t("early_career_payments.questions.nqt_in_academic_year_after_itt"), "Yes", "nqt-in-academic-year-after-itt"]
+      [I18n.t("early_career_payments.questions.nqt_in_academic_year_after_itt"), "Yes", "nqt-in-academic-year-after-itt"],
+      [I18n.t("early_career_payments.questions.postgraduate_itt_or_undergraduate_itt_course"), "Postgraduate", "postgraduate-itt-or-undergraduate-itt-course"]
     ]
 
     expect(presenter.answers).to eq(expected_answers)